pyproject.toml
[tool.poetry]
name = "django-image-downloader"
version = "0.1.0"
description = ""
authors = ["mintydevdaz <mintydevdaz@proton.me>"]
readme = "README.md"
packages = [{include = "django_image_downloader"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
validators = "^0.20.0"
httpx = "^0.24.1"
selectolax = "^0.3.14"
playwright = "^1.35.0"
pillow = "^10.0.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
